WB-274: added member file download and new file icons

+++ src/app/features/scouts/components/document-list/document-list.component.ts
@@ -1,13 +1,15 @@
 import {Component, inject, input, output, viewChild} from '@angular/core';
 import {MemberFile} from "../../models/member.model";
 import {TableModule} from "primeng/table";
-import {DatePipe, NgClass} from "@angular/common";
+import {DatePipe, NgClass, NgOptimizedImage} from "@angular/common";
 import {FileUtils} from "../../../../shared/util/file.utils";
 import {Tag} from "primeng/tag";
 import {FileUpload, FileUploadHandlerEvent} from "primeng/fileupload";
 import {maxFileUploadByteSize} from "../../../../shared/constant";
-import {finalize, forkJoin, noop, Observable} from "rxjs";
+import {finalize, forkJoin, Observable} from "rxjs";
 import {AlertService} from "../../../../shared/services/alert-service.service";
+import {ScoutService} from "../../services/scout.service";
+import {ConfirmationService} from "primeng/api";
 
 @Component({
   selector: 'app-document-list',
@@ -16,12 +18,19 @@ import {AlertService} from "../../../../shared/services/alert-service.service";
     Tag,
     DatePipe,
     FileUpload,
-    NgClass
+    NgClass,
+    NgOptimizedImage
   ],
   templateUrl: './document-list.component.html',
   styleUrl: './document-list.component.scss'
 })
 export class DocumentListComponent {
+
+  protected alertService = inject(AlertService);
+  protected scoutService = inject(ScoutService);
+  protected confirmationService = inject(ConfirmationService);
+
+  protected readonly maxFileUploadByteSize = maxFileUploadByteSize;
   protected readonly FileUtils = FileUtils;
 
   documents = input.required<MemberFile[]>();
@@ -30,17 +39,12 @@ export class DocumentListComponent {
   canEdit = input<boolean>(true);
   protected onUpload = output<MemberFile[]>();
 
-  protected loading: boolean = false;
-  protected alertService = inject(AlertService);
-
   private readonly uploader = viewChild.required(FileUpload);
-  protected readonly maxFileUploadByteSize = maxFileUploadByteSize;
+
+  protected loading: boolean = false;
 
   protected getFileIcon(file: MemberFile): string {
-    if (file.name.endsWith('.pdf')) {
-      return "assets/pdf.png";
-    }
-    return "assets/text.png";
+    return FileUtils.getFileIcon(file.mimeType);
   }
 
   get canEditDocuments() {
@@ -68,15 +72,20 @@ export class DocumentListComponent {
 
   protected deleteFile(fileToDelete: MemberFile, index: number) {
     if (this.loading) return;
-    this.loading = true;
-    this.deletePetition()!(fileToDelete.id)
-      .pipe(finalize(() => this.loading = false))
-      .subscribe({
-        next: () => {
-          this.documents().splice(index, 1);
-          this.alertService.sendBasicSuccessMessage("Documento eliminado con éxito");
-        }
-      });
+    this.confirmationService.confirm({
+      message: "¿Desea borrar este archivo? Esta acción no se podrá deshacer",
+      accept: () => {
+        this.loading = true;
+        this.deletePetition()!(fileToDelete.id)
+          .pipe(finalize(() => this.loading = false))
+          .subscribe({
+            next: () => {
+              this.documents().splice(index, 1);
+              this.alertService.sendBasicSuccessMessage("Documento eliminado con éxito");
+            }
+          });
+      }
+    });
   }
 
   protected openUploader() {
@@ -84,5 +93,14 @@ export class DocumentListComponent {
     this.uploader().el.nativeElement.getElementsByClassName("p-fileupload-choose-button")[0].click();
   }
 
-  protected readonly noop = noop;
+  protected downloadFile(fileId: number) {
+    this.loading = true;
+    this.scoutService.getMemberFile(fileId)
+      .pipe(finalize(() => this.loading = false))
+      .subscribe(res => FileUtils.downloadFile(res));
+  }
+
+  protected openFile(fileId: number) {
+    FileUtils.openFile(this.scoutService.getMemberFile(fileId));
+  }
 }
